git-r3.eclass: Remove unnecessary HEAD->branch mapping code, #524146
Remove the code responsible for recognizing which branch HEAD pointed
out to since it was unsafe and unnecessarily complex. A proper match is
not really necessary since all operations can be safely performed on
an opaque 'HEAD' (or rather refs/git-r3/HEAD since fetching to HEAD
directly breaks git a bit).