Add MySQL packaging changes announcement
The purpose of this news item is to announce to users the USE flag
changes and that a server package should be added to the world file
if one is required locally.  Future dependencies may not pull the
server automatically as it was in the past.
Also includes a suggestion to resolve potential blockers by setting
the new client-libs USE on a server package