microcode-ctl: stop installing the init script
Updating microcode on the fly is dangerous as it can modify the set of
valid instructions.  An active example of this is Intel's TSX insns --
the latest microcode push disables the insn on newer CPUs and causes
SIGILL when you try to use it.  But if you test for the insn before the
microcode is updated, it will execute fine.  For daemons that launched
before the update, they'll find the flag works, and then crash later on
when the insn no longer exists.

Thus the only safe way to update microcode is at boot time via a builtin
initramfs.  Details on this operation can be found in #528712#41.