sys-libs/ncurses: use libgpm.so.1 all the time #544828
Since the gpm code relies on running ldd on the linked file in order to
calculate the SONAME, and we're passing this in as a configure flag some
of the time, just change the code to always pass it in.  This stabilizes
behavior across cross-compiling/multilib/ldd output/etc...  Relying on
ldd output in general is bad juju.