dev-lang/go-bootstrap: avoid stripping test data files
Go includes a number of pre-built object files for testing. Since they
are not actually executed and may not even be native to the local system
stripping them is meaningless. This syncs up STRIP_MASK with the same
list used by the latest go ebuilds.

One file, src/debug/dwarf/testdata/typedef.elf, includes a build-id
which causes a conflict with the go-1.4 when FEATURES=splitdebug is
enabled.

Fixes bug #549598