app-admin/logstash-forwarder: Allow to specify user/group for service
logstash-forwarder is running as logstash user in Gentoo by default.  It
might be desired to run it as other user who has access to log files,
presumably as root.  The other solution for accessing logs would be
adding logstash user to groups as which services are run and make logs
readable by that group.  It is also possible in some cases to change
group for log files, e.g. in syslog-ng.

Thanks to Tomas Mozes <hydrapolic@gmail.com> for submitting patch.

Gentoo-Bug: 557462

Package-Manager: portage-2.2.20.1