"globally" package.use.mask www-client/chromium[widevine]
...by means of profiles/base, and then un-package.use.mask
it on those platforms where it's supported.

This allows us to drop the sexy-but-probably-depsolvingly-expensive
REQUIRED_USE logic from the chromium ebuild.

I'm surprised there is only one example of this in the tree.  Is there
some best-practice to effect a USE flag that's whitelisted on a few
platforms and masked everywhere else?  Surely it's not to use.mask our
local flag globally?  That's crazy-talk.

Package-Manager: portage-2.2.20.1
Signed-off-by: Gregory M. Turner <gmt@be-evil.net>