php-ext-source-r2.eclass: add prefix support.
The php-ext-source-r2 eclass sets a bunch of environment variables of
the form,

  PHPIZE="/usr/${libdir}/${1}/bin/phpize"

and also tests for the existence of an INI file using absolute
paths. To accomodate a prefix environment, all of these paths need to
be prefixed with $EPREFIX. This commit adds the prefix to all absolute
paths in the eclass. We then need to strip $EPREFIX from one
invocation of `insinto` which handles $EPREFIX in its own way.

This has been tested on =dev-php/pecl-imagick-3.3.0_rc2 in a prefix
environment.

Bug: 371985
Bug: 481410