php-ext-source-r2.eclass: fix WANT_AUTO{CONF,MAKE} warnings.
This is a real fix for the warnings like,

  ac-wrapper: warning: invalid WANT_AUTOCONF 'latest'; ignoring.

reported in bugs #329071 and #549268. In the first of the two, vapier
suggested running $PHPIZE through the autotools_run_tool wrapper, but
this caused builds to fail. In bug #329071, that's where the story
ends.

It turns out, the build failures were due to $PHPIZE containing an
absolute path. Vapier fixed the path handling in autotools.eclass
(commit ecd9954), so now we are able to fix this properly in
php-ext-source-r2.eclass.

Bug: 329071
Bug: 549268