firefox,thunderbird: Ensure $SHELL gets set to something
Zero_Chaos noticed when building firefox within catalyst that failures would
occur due to the build system not finding the correct shell.  It was determined
this was due to catalyst not setting SHELL in the environment, which resulted in
the firefox build system being passed SHELL="" and causing the failure.

This commit ensures SHELL will never be empty by hard-coding /bin/bash (EPREFIX'd
when appropriate) if there is no SHELL specified in the environment.

Changed ebuilds:
www-client/firefox-38.2.1
www-client/firefox-40.0.3
mail-client/thunderbird-38.2.0

Package-Manager: portage-2.2.20.1