dev-java/oracle-jre-bin: Remove PostVMInitHook to stop usage tracker
Deleting the usage tracker resulted in exceptions being silently
raised in a thread and this caused havoc with debuggers. Deleting the
hook that calls it instead is a smaller change and avoids both
problems.

It is possible that Oracle may use this hook for something else
later. I have therefore added a JAVA_PKG_STRICT check to display its
contents on each merge so that we can keep an eye on it.

See bug #559936 for more details.

Package-Manager: portage-2.2.20.1