sys-devel/libtool: force disable help2man w/releases #556512
We want to make sure that help2man isn't run inadvertently (since we
explicitly do not depend on it w/releases), so set HELP2MAN to `false`
so it hard fails immediately.

We also fix the build-aux/ltmain.in patch as it triggers rebuilding of
build-aux/ltmain.sh which triggers rebuilding of doc/libtool.1 which
requires `help2man`.