mail-filter/pypolicyd-spf: remove python2 support to avoid ipaddress conflict.
Current versions of pypolicyd-spf support both python2 and
python3. With python2, the dev-python/ipaddr library provides support
for IP address parsing. And with python3 (>= 3.3), the built-in
ipaddress module is used. In fact, this determination is made based on
the presence (or absence) of the ipaddress module.

So far this has been fine. But recently, dev-python/ipaddress was
introduced to backport the python3 ipaddress module to python2. Now,
when dev-python/ipaddress is detected, pypolicyd-spf (incorrectly)
assumes the python3 implementation. This is bug #559364 by Robert
Trace.

Upstream will be python3-only in the next release. They have suggested
that we drop support for python2, and that's what this revision
does. Since this can cause mail to be rejected or delayed, the new
revision replaces the original.

Bug: 559364

Package-Manager: portage-2.2.20.1