dev-java/jnr-unixsocket: Bump Java version to 7 as jnr-enxio also has
This relates to bug #561352. jnr-unixsocket builds with Java 6 but it
should have a Java version at least as high as its dependencies.

I have also changed the jnr-ffi SLOT to match the one used by
jnr-enxio, otherwise you end up with two conflicting versions on the
classpath. jruby (a revdep) also uses the older version but it is
masked and totally broken anyway.

This demands a revbump but should still be safe enough to commit
straight to stable.

Package-Manager: portage-2.2.20.1