dev-php/pecl-redis: revision bump to fix incorrect DOCS.
Update the DOCS array to contain accurate filenames. Due to a bug in
php-ext-source-r2.eclass, it currently is not an error when DOCS
contains bad values. Before that can change, we need to fix the
existing ebuilds.

As part of the revbump, a few other cosmetic issues were fixed.

Package-Manager: portage-2.2.20.1