profiles/use.stable.mask: mask ruby_targets_ruby22
ruby22 is currently in testing, and starting with this
version we have adopted the strategy of the python
team to mask testing-only versions this way.

This facilitates the stable process for ruby packages
and enables us to move ruby22 to stable at some
point in the future with good coverage of packages
right at the start.