dev-java/bprov: Add back old ebuilds for version 1.50 and 1.52 until SLOTing is fixed across the tree.
Currently, both ebuilds have two different SLOTs (0 and 1.52) which is
inconsistent. We should decide on which SLOT we keep and mend ebuilds across
the tree who use the wrong SLOT.

I have been breaking the tree a couple of times this morning due to this
mismatch. This commit should quiet down the CI system now.

Package-Manager: portage-2.2.20.1
RepoMan-Options: --force
Signed-off-by: Patrice Clement <monsieurp@gentoo.org>