metadata: add slots element
There seems to be some general confusion about specific package SLOTs
and their meaning, since there can be several naming schemes applied
and documentation is either non-existent or is inside the ebuild via
comments.
Because of that it should be part of metadata.xml.

An example use case for media-libs/libpng would be:
<slots>
    <slot name="0">For building against. This is the only slot
        that provides headers and command line tools.</slot>
    <slot name="1.2">For binary compatibility, provides libpng12.so.0.</slot>
    <slot name="1.5">For binary compatibility, provides libpng15.so.15.</slot>
    <subslots>Represent ABI compatibility for libpng.so.</subslots>
</slots>

For packages like x11-libs/wxGTK one could write:
<slots>
    <slot name="*">Major versions.</slot>
</slots>