dev-libs/libintl: drop USE=nls flag
The related commit in gettext (030b323567f9546c35d9cf1182f36b2b54e08)
was wrong -- the internal configure logic skips the libintl flag when
nls support is disabled.  It doesn't matter there now as libintl has
been dropped, but it still matters here as w/out the flag, no library
is built.  So drop the flag, hardcode the option, and add a comment
to keep from screwing it up in the future.