dev-cpp/antlr-cpp: New package, split from dev-java/antlr
The old ANTLR 2 ebuild tried to cram all the runtimes together, which
was messy. The C++ runtime does not need any Java components at build
time or runtime so it makes sense to split this out. The C++ runtime
for ANTLR 4 is maintained separately by upstream anyway.

This new package does a better job of building the shared library with
the help of libtool. It also supports multilib.

Package-Manager: portage-2.2.20.1