app-portage/gentoolkit: fix eclean -n -d, #427020
Running "eclean-pkg -n -d" keeps binpkgs even without a corresponding ebuild,
which is not what I expected from the documentation, which says:

  "Somewhere  in the middle, adding the --package-names option when using
  --destructive will protect files corresponding to all existing versions of
  installed packages."

I interpreted "existing versions of installed packages" to mean versions which
are installed and which also exist in the portage tree.  However, what eclean
does in findPackages() (with "-d -n") is the following three things:

- check whether the cat/pkg-ver is listed in an exclusion file,
- check whether the cat/pkg-ver exists in the vartree, meaning it only filters
  out the newest version, and
- check whether the cat/pkg exists.

The first point is irrelevant to this bug, and the second one is clear, since
that's what "-d" by itself is supposed to do. However, the third point means
that "eclean -d -n" will never delete binpkgs that belong to an installed
cat/pkg.

To fix this, it is necessary to change findPackages() to also check that the
package version exists, i.e., to change the relevant conditional from

  if portage.cpv_getkey(cpv) in cp_all:

to

  if portage.cpv_getkey(cpv) in cp_all and port_dbapi.cpv_exists(cpv):

which is exactly what the proposed patch does.

With this patch applied "eclean -d -n -p" reports almost 1 GB of packages to
remove, including versions of samba and udev that don't exist anymore. The
output appears to be the same as that of "eclean -p", save for a version of
gentoo-sources that doesn't exist anymore, but that I still have installed
(which I think verifies that the logic is correct).

My motivation for looking into this was that old firefox binpkgs were not being
deleted, despite the fact that the corresponding ebuilds did not exist anymore.
Furthermore, my weekly eclean cronjob kept reporting "Your packages directory
was already clean.". In contrast, "eclean-dist -n -d" has been working fine so
far and has, e.g., deleted all old firefox distfiles.

[ Note that this patch and the corresponding bug report are now over two years
  old, and that I have been running this patch on both of my systems for just
  as long.  Apparently there is a better way to do it (see the comment by Brian
  Dolbec in the bug report), but that hasn't happened yet, and it is unclear to
  me when it will. ]

Gentoo-bug: 427020
Signed-off-by: Marc Joliet <marcec@gmx.de>