x11-wm/plwm: take maintainership, drop to ~arch
This commit cleans up the versions that used the old python.eclass,
and also drops stable keywords as the age of the upstream package
combined with the fact that I don't use it means to me that it isn't
a good stable candidate.  Please file a bug if there is a particular
need from end users to have this package stabilized.

Package-Manager: portage-2.2.20.1