wxwidgets.eclass: Define -DNDEBUG when building against wxGTK:3.0
In wxGTK 3.0 debugging support is always enabled in the library and
it's up to packages to state if they want to use it or not.  When
enabled this makes warning dialogs pop up any time an assertion
is encountered, which isn't very user friendly.  This disables that
behavior by adding -DNDEBUG to CPPFLAGS unless the package has a
debug USE flag and it is enabled.

Gentoo-Bug: https://bugs.gentoo.org/562480