dev-lang/php: slot openssl dependency.
The dev-libs/openssl package now has two slots. The SLOT=0.9.8 is for
binary compatibility, and SLOT=0 is what most packages should depend
on. Adding ":0" to our openssl dependency ensures that we get the
correct slot, and silences a repoman warning.

The lower bound on openssl was also removed since all versions in the
tree satisfied it.

Package-Manager: portage-2.2.20.1