profiles: restore sys-apps/attr and dev-python/pyxattr for musl stage1
Commit 824d0647 add USE=xattr to the base profile and removed the flag
from other profiles as well as dropped attr and pyxattr form stage1
builds.  This, however, breaks at least the arm musl build because
at stage3 portage is missing pyxattr or get/setfattr.  One approach
would be to add BOOTSTRAP_USE="xattr", the other is to just make
sure attr and pyxattr are in stage1->stage2 for the stage3 build.
We choose the later for now.