dev-lang/php: revbump php-5.6.15 to fix a segfault.
The segfault issue was reported in bug #564690 by Zoltán
Halassy. There is an upstream fix that we've added as a patch. The
problem/fix have both been confirmed.

Since this requires a revbump, the sys-libs/db dependency has been
fixed along with it according to bug #521222. It still remains to
address that issue for the php-7.x series.

Gentoo-Bug: 521222
Gentoo-Bug: 564690

Package-Manager: portage-2.2.20.1