dev-java/javacup: Rewrite for version bump to 0.11b_p20151001
Dropped the beta label as the previous version was not really a beta.

No longer has a system-jflex USE flag now that jflex has javacup as a
PDEPEND. This is better as users were unlikely to ever set this flag
and a system version of javacup will be used automatically when it
becomes available.

~arm keyword restored and ~ppc64 keyword temporarily dropped as Java 7
is required now.

Package-Manager: portage-2.2.20.1