profiles: update sourceforge mirror group
This replaces all of the existing mirrors with the bouncer
download.sourceforge.net, which automatically points to the best mirror.

SourceForge mirrors seem to change often (many that were listed here no longer
exist) and are not all guaranteed to carry every file. This made it sometimes
difficult or even impossible to fetch files.

Gentoo-bug: 564330