dev-lang/php: don't depend on versions of sys-libs/db that won't be detected.
In bug #564824, Robert Förster noticed that there's an issue building
PHP with both sys-libs/db-5.3 and sys-libs/db-6.0 installed. The root
cause of the failure is that PHP attempts to autodetect sys-libs/db,
but doesn't specifically look for db-5.3. When db-5.3 is not found,
it falls back to /usr/include/db.h from version 6.0 and that version
is incompatible.

A similar issue exists with older 4.x versions, but not with db-4.5
through db-5.1. A fix has been sent upstream for db-5.3. In the
meantime, we remove that and the older 4.x versions from $DEPEND.

Gentoo-Bug: 564824

Package-Manager: portage-2.2.20.1