sys-apps/busybox: move USE=static default to the profile #468580
Since busybox is pulled in via the linux profile (as we use it as the
rescue shell there), we should have the static USE flag set via the
profile too rather than the ebuild.  This way we apply it only to the
subset of profiles where it makes sense.

For the live ebuild, we also switch to REQUIRED_USE since package.use
is able to override make.defaults (which is where USE=pam is set).  We
can see how users react to this behavior before pulling it back into
the released versions.