www-client/dooble: new package, "a secure web browser"
This is a rewrite of the www-client/dooble package from Sabayon's
for-gentoo repository:

https://github.com/Sabayon/for-gentoo/tree/master/www-client/dooble

I have not audited the package to verify the security claims, but a
cursory peek at how it restricts possible exploit vectors has been
promising. I would like this to be scrutinized by others, but I feel
good enough about this that I am committing it to ~amd64 for our users
to evaluate.

At the moment, I am only committing the Qt4 version, which is the 1.x
series. The 2.x series is the Qt5 version. This probably should be
changed to be slotted when that is committed.

Package-Manager: portage-2.2.20.1