app-emulation/wine: Version bump to 1.7.54.  Drop default prelink flag.
Wine 1.7.54 adds a patch to use the system linker if possible before
attempting to use prelink.  As such, dropping default status for prelink.
Hardened users are unable to take advantage of this and may still want to
unmask and use the prelink flag, though this isn't well supported.

Package-Manager: portage-2.2.23