dev-java/antlr: Version bump for SLOT 3, should fix #567560
I initially wanted to add 3.3 but despite it being less disruptive
than 3.4, I found compatibility issues with the revdeps. We will
hopefully never need 3.3.

This is a total rewrite from the previous version, which uses the new
simplified antlr-2 ebuild and adds a very large test suite that is
passing with a little patching.

Package-Manager: portage-2.2.26