dev-java/antlr: Version bump to 4.5.1 (closes bug #545384)
This is a rewrite as the 4.4 ebuild bundled a lot of prebuilt stuff
and didn't include any tests. gunit support has been dropped because
no one uses it and it is not worth the hassle.

Nothing in the tree truly needs at present this as the existing
dependencies are bogus but no doubt something will before long.

Package-Manager: portage-2.2.26