dev-lang/python: Ensure python-config & .pc ABI symlinks, idle wrapping
Fix the pythonX.Y-config symlink creation to be only applied when
ABIFLAGS are not empty, and add additional python-X.Ym.pc symlink for
pkg-config. Make both official as they are created as part of regular
upstream 'make install' (but not 'make altinstall' used by Gentoo).

Install python-exec wrapping symlink for the idle executable when USE=tk
is enabled.