media-video/mpv: adjust required libass USE flags wrt bug 562394
mpv uses libass to render subtitles and enca to detect subtitle
encoding. For mpv to support enca, mpv does not need to have
libass[enca]. This helps to avoid problems with >=libass-0.13.0,
which dropped 'enca' USE.

Also enable harfbuzz USE to support subtitles for complex languages.

See [0] as a reference for both of the above.

[0]: https://github.com/mpv-player/mpv/blob/master/DOCS/man/options.rst

Gentoo-Bug: 562394

Package-Manager: portage-2.2.24