x11-libs/gtk+:  subslot gtk+:3
The goal of this subslot is to allow other packages to determine what minor
version of gtk+:3 is installed so that they can operate accordingly.

This functionality is extremely desirable as there are often incompatible
changes between gtk+:3 branches.  This functionality enables an ebuild to
make decisions based on the branch via has_version or best_version queries
during phases.

The best way to take advantage of this functionality is
DEPEND="x11-libs/gtk+:3=" or similar. This has the obvious issue that there
is no guarantee that the subslot will be consistent throughout the phases
of the ebuild, however, because of the subslot rebuild operator, the
package manager should force rebuilding ebuild in question.

Package-Manager: portage-2.2.25