btrfs: add static(-libs) USE falgs
btrfs is slowly becoming the new main fs. For certain situations, it may
be preferable to have statically linked binaries. To facilitate this,
the static-libs flag has been added.

Additionally, all the btrfs utilities can be statically compiled (for
use in custom initramfs for example). To this end, the static flag has
been added.

To get everything out of btrfs, sys-apps/util-linux should also be
emerged with static-libs for libuuid and libblkid. The btrfs Makefile
does auto detect the presence of these (static) libraries so as such it
is not a hard requirement.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>