app-eselect/eselect-php: new revision with backwards-compatible apache config.
If the user has an unmodified 70_mod_php5.conf, the previous revision
will allow that file to be removed. This can break PHP support in
Apache before the user notices that he needs to run eselect-php and
update his Apache -DPHP5 line.

In this revision we supply the latest copy of 70_mod_php5.conf for
backwards compatibility. There is now a notice at the top of that file
informing users to upgrade their configuration.

Package-Manager: portage-2.2.24