dev-lang/php: update the apache2 post-install notification about mod_php.
Newer versions of app-eselect/eselect-php require some apache
configuration changes. This configuration is mentioned in the
dev-lang/php pkg_postinst() phase, which isn't ideal, but is what it
is. Since php:7.0 requires the new eselect-php, the pkg_postinst() for
php:7.0 has been updated to reflect the configuration changes.

Thanks to Tomáš Mózes for pointing this out.

Gentoo-Bug: 552156

Package-Manager: portage-2.2.24