app-text/ghostscript-gpl: fix zlib include path usage
The ghostscript-gpl-9.12-sys-zlib.patch attempts to change the include
path from the local copy in the src to the common includedir flag set
by configure flags.  This generally works for native builds, but breaks
when cross-compiling.  Drop the patch and set the zlib src dir to the
temp dir instead.