dev-java/gnu-classpath: Build with regular javac instead of ECJ
I believe GNU Classpath favoured ECJ because it was the only FOSS
compiler capable of building it at the time. OpenJDK is now preferable
and the statement in configure.ac that not using ECJ results in
com.sun.tools.javac not being included doesn't hold true.

The simplification is nice but this also avoids a race condition
occuring with ECJ on ppc. This is almost certainly a CACAO issue. See
http://icedtea.classpath.org/bugzilla/show_bug.cgi?id=2782.

Package-Manager: portage-2.2.26