sys-libs/ldb: resolve automagic header usage on libbsd by RDEPENDing on it
The lib/replace code configures and depends on bsd/* headers depending solely on its
existence; as libbsd is a non-optional dep for samba-4.2 and newer anyways, ensure
it is installed now with matching MULTILIB_USDEP on non-fbsd to avoid the potential
headers mis-matches between bsd/*.h and other unistd.h or types.h on the system.

This should address bug 571622.

Package-Manager: portage-2.2.26