app-backup/backuppc: add version 3.3.1 (masked) with apache-2.4 support.
This new version adds apache-2.4 support to backuppc, but at the cost
of requiring www-apache/mpm_itk and a new mod_perl supporting
apache-2.4. The former is available and less onerous than the current
requirement of a separate web server. However, the only mod_perl
supporting apache-2.4 at the moment is masked, so the new version of
backuppc is masked as well.

The configuration for the new version should be more familiar to
users: it is enabled by passing -D BACKUPPC to apache2.

Gentoo-Bug: 532718
Gentoo-Bug: 541312

Package-Manager: portage-2.2.26