dev-lang/php: revision the 5.6 and 7.0 series to add pdo-firebird support.
PHP has two Firebird drivers -- the legacy "interbase" driver and the
new pdo-firebird driver. We have always supported the legacy driver
with USE=firebird, but were missing the pdo-firebird configure flag
when USE="pdo firebird" was set. Similar tricks were in place for
e.g. sqlite and oci8, but Firebird was forgotten.

Dmitry A. Bakshaev was nice enough to point this out in bug 538226,
and even supplied a patch. Testing shows that both drivers work in
php:5.6, but that only the PDO driver works in php:7.0. That's not
a regression; this commit adds the only driver that works in php:7.0.

Gentoo-Bug: 538226
Suggested-By: Dmitry A. Bakshaev

Package-Manager: portage-2.2.26