dev-lang/php: enforce what was an elog warning with REQUIRED_USE.
The interactive "php -a" command-line interface will hang unless PHP
is built with USE=readline. This used to be an elog warning, but was
rather easy to overlook and mysterious to those who hit the bug. We
already have REQUIRED_USE for our other flags, so it makes sense to
enforce the cli dependency on readline there.

Package-Manager: portage-2.2.26