dev-java/piccolo2d: Remove old
This drops the only stable version but it doesn't build at all now
(was sun.dc part of the JDK before?) and it currently has no revdeps.

The new version should probably have been put in a different SLOT but
that doesn't matter now that the old version is going.

Package-Manager: portage-2.2.27