www-client/firefox-44.0: drop EGL use flag
EGL support in firefox-44 is non-optional, but it is also not the default GL provider; this
flag could set EGL to be the default GL provider instead of GLX, but there are both compile time
and link time issues with attempting to do this.  Removing the flag until the build-time situation
has improved.

Bug: https://bugs.gentoo.org/571180

Package-Manager: portage-2.2.26