php-pear-r1.eclass: add support for EAPI=6 and clean up the EAPI check.
The dev-lang/php dependency in php-pear-r1.eclass is calculated based
on the EAPI. In newer EAPIs, we specify the "any slot" operator to
avoid repoman warnings. Previously, the "any slot" operator was added
only for EAPI=5; this commit adds it for EAPI=6.

In addition, EAPIs 0, 1, 2, 3, and 4 are named explicitly in the case
statement. The default "*" case now dies with a warning that the EAPI
is unsupported.