xfce-base/xfce4-session: screensaver fixes
Add x11-misc/light-locker as one of the possible backends for locking
the screen. Add the corresponding command to the xflock4 script.

Remove gnome-extra/gnome-screensaver as one of the possible backends for
locking the screen. This silences repoman. Gnome 3 does not use it and
Gnome 2 no longer exists in portage, thus the package has disappeared.
Leave the corresponding command in the xflock4 script just in case some
of our users are still using gnome-screensaver from an overlay.