dev-java/typesafe-config: Resolve bug #559748 once and for all
typesafe-config-1.3.0 has a stable revdep (mchange-commons) but
requires Java 8, which hasn't been stabilised yet. An ebuild depending
on Java 6 was erroneously added to deal with this. We're not quite
ready to stabilise Java 8 but we can't drop keywords either so I've
added typesafe-config-1.2.1, which only requires Java 6. java-apicheck
highlighted some differences but I've kept the same SLOT because
mchange-commons works with either.

Package-Manager: portage-2.2.27