sys-devel/bison: avoid bison.info regeneration #574492
The build logic generates a timestamp from the info pages, but does so
with a timezone aware function, and then truncates it to granularity of
days.  If your TZ happens to wrap the timezone over to a diff day, then
the build logic updates some files and runs makeinfo.

Clobber the info timestamp so it's newer than the sources and the TZ
state doesn't work.