www-apache/mod_caucho: Remove old
This drops the only stable version but there are no revdeps and we
need to clear this for the removal of Java 6. I recommend not marking
it stable again until someone has done more than just build it. I
don't use Resin or Apache and I'm not entirely confident that it would
work.

Package-Manager: portage-2.2.27